replaced ACCESS_TOKEN and WPCOM_USERNAME preferences by a call to the default account
