+++ WordPress/src/androidTest/java/org/wordpress/android/mocks/OAuthAuthenticatorEmptyMock.java
@@ -1,15 +1,11 @@
 package org.wordpress.android.mocks;
 
-import android.content.SharedPreferences;
-import android.preference.PreferenceManager;
-
-import org.wordpress.android.WordPress;
 import org.wordpress.android.networking.AuthenticatorRequest;
 import org.wordpress.android.networking.OAuthAuthenticator;
+import org.wordpress.android.util.AccountHelper;
 
 public class OAuthAuthenticatorEmptyMock extends OAuthAuthenticator {
     public void authenticate(AuthenticatorRequest request) {
-        final SharedPreferences settings = PreferenceManager.getDefaultSharedPreferences(WordPress.getContext());
-        settings.edit().putString(WordPress.ACCESS_TOKEN_PREFERENCE, "dead-parrot").commit();
+        AccountHelper.getDefaultAccount().setAccessToken("dead-parrot");
     }
 }
