+++ WordPress/src/main/java/org/wordpress/android/datasets/AccountTable.java
@@ -22,7 +22,6 @@ public class AccountTable {
     public static void createTables(SQLiteDatabase db) {
         db.execSQL("CREATE TABLE " + ACCOUNT_TABLE + " ("
                 + "	local_id                INTEGER PRIMARY KEY DEFAULT 0,"
-                + "	is_logged_in            INTEGER DEFAULT 0,"
                 + "	is_wordpress_com_user   INTEGER DEFAULT 0,"
                 + "	user_name	            TEXT,"
                 + "	user_id	                INTEGER DEFAULT 0,"
@@ -47,7 +46,6 @@ public class AccountTable {
         ContentValues values = new ContentValues();
         // we only support one wpcom user at the moment: local_id is always 0
         values.put("local_id", 0);
-        values.put("is_logged_in", account.isSignedIn());
         values.put("is_wordpress_com_user", account.isWordPressComUser());
         values.put("user_name", account.getUserName());
         values.put("user_id", account.getUserId());
@@ -73,7 +71,6 @@ public class AccountTable {
 
         try {
             if (c.moveToFirst()) {
-                account.setIsLoggedIn(SqlUtils.sqlToBool(c.getInt(c.getColumnIndex("is_logged_in"))));
                 account.setIsWordPressComUser(SqlUtils.sqlToBool(c.getInt(c.getColumnIndex("is_wordpress_com_user"))));
                 account.setUserName(c.getString(c.getColumnIndex("user_name")));
                 account.setUserId(c.getLong(c.getColumnIndex("user_id")));
