+++ WordPress/src/main/java/org/wordpress/android/networking/OAuthAuthenticator.java
@@ -1,19 +1,15 @@
 package org.wordpress.android.networking;
 
-import android.content.SharedPreferences;
-import android.preference.PreferenceManager;
-
 import org.wordpress.android.WordPress;
 import org.wordpress.android.models.Blog;
+import org.wordpress.android.util.AccountHelper;
 import org.wordpress.android.util.StringUtils;
 
 public class OAuthAuthenticator implements Authenticator {
     @Override
     public void authenticate(final AuthenticatorRequest request) {
         String siteId = request.getSiteId();
-
-        SharedPreferences settings = PreferenceManager.getDefaultSharedPreferences(WordPress.getContext());
-        String token = settings.getString(WordPress.ACCESS_TOKEN_PREFERENCE, null);
+        String token = AccountHelper.getDefaultAccount().getAccessToken();
 
         if (siteId != null) {
             // Get the token for a Jetpack site if needed
