+++ WordPress/src/main/java/org/wordpress/android/ui/ShareIntentReceiverActivity.java
@@ -20,6 +20,7 @@ import org.wordpress.android.models.Blog;
 import org.wordpress.android.ui.accounts.SignInActivity;
 import org.wordpress.android.ui.media.MediaBrowserActivity;
 import org.wordpress.android.ui.posts.EditPostActivity;
+import org.wordpress.android.util.AccountHelper;
 import org.wordpress.android.util.BlogUtils;
 import org.wordpress.android.util.ToastUtils;
 
@@ -111,7 +112,7 @@ public class ShareIntentReceiverActivity extends ActionBarActivity implements On
     private void finishIfNoVisibleBlogs() {
         // If not signed in, then ask to sign in, else inform the user to set at least one blog
         // visible
-        if (!WordPress.isSignedIn()) {
+        if (!AccountHelper.getDefaultAccount().isSignedIn()) {
             ToastUtils.showToast(getBaseContext(), R.string.no_account, ToastUtils.Duration.LONG);
             startActivity(new Intent(this, SignInActivity.class));
             finish();
