+++ WordPress/src/main/java/org/wordpress/android/ui/accounts/helpers/UpdateBlogListTask.java
@@ -1,13 +1,11 @@
 package org.wordpress.android.ui.accounts.helpers;
 
 import android.content.Context;
-import android.content.SharedPreferences;
 import android.os.AsyncTask;
-import android.preference.PreferenceManager;
 
-import org.wordpress.android.WordPress;
 import org.wordpress.android.ui.accounts.BlogUtils;
 import org.wordpress.android.ui.accounts.helpers.FetchBlogListAbstract.Callback;
+import org.wordpress.android.util.AccountHelper;
 import org.wordpress.android.util.AppLog;
 import org.wordpress.android.util.AppLog.T;
 import org.wordpress.android.util.CoreEvents;
@@ -36,8 +34,7 @@ public class UpdateBlogListTask extends AsyncTask<Void, Void, List<Map<String, O
 
     @Override
     protected List<Map<String, Object>> doInBackground(Void... args) {
-        SharedPreferences settings = PreferenceManager.getDefaultSharedPreferences(mContext);
-        final String username = settings.getString(WordPress.WPCOM_USERNAME_PREFERENCE, null);
+        final String username = AccountHelper.getDefaultAccount().getUserName();
         final CountDownLatch countDownLatch = new CountDownLatch(1);
         FetchBlogListWPCom fetchBlogList = new FetchBlogListWPCom();
         fetchBlogList.execute(new Callback() {
