+++ WordPress/src/main/java/org/wordpress/android/ui/comments/CommentDetailFragment.java
@@ -55,6 +55,7 @@ import org.wordpress.android.ui.suggestion.adapters.SuggestionAdapter;
 import org.wordpress.android.ui.suggestion.service.SuggestionEvents;
 import org.wordpress.android.ui.suggestion.util.SuggestionServiceConnectionManager;
 import org.wordpress.android.ui.suggestion.util.SuggestionUtils;
+import org.wordpress.android.util.AccountHelper;
 import org.wordpress.android.util.AniUtils;
 import org.wordpress.android.util.AppLog;
 import org.wordpress.android.util.AppLog.T;
@@ -224,7 +225,8 @@ public class CommentDetailFragment extends Fragment implements NotificationFragm
         mLayoutReply = (ViewGroup) view.findViewById(R.id.layout_comment_box);
         mEditReply = (SuggestionAutoCompleteText) mLayoutReply.findViewById(R.id.edit_comment);
         mEditReply.getAutoSaveTextHelper().setUniqueId(String.format("%s%d%d",
-                WordPress.getLoggedInUsername(WordPress.getCurrentBlog()), getRemoteBlogId(), getCommentId()));
+                AccountHelper.getDefaultAccount().getCurrentUsername(WordPress.getCurrentBlog()),
+                getRemoteBlogId(), getCommentId()));
 
         mImgSubmitReply = (ImageView) mLayoutReply.findViewById(R.id.image_post_comment);
 
@@ -623,7 +625,7 @@ public class CommentDetailFragment extends Fragment implements NotificationFragm
         // the post this comment is on can only be requested if this is a .com blog or a
         // jetpack-enabled self-hosted blog, and we have valid .com credentials
         boolean isDotComOrJetpack = WordPress.wpDB.isRemoteBlogIdDotComOrJetpack(mRemoteBlogId);
-        boolean canRequestPost = isDotComOrJetpack && WordPress.hasDotComToken();
+        boolean canRequestPost = isDotComOrJetpack && AccountHelper.getDefaultAccount().hasAccessToken();
 
         final String title;
         final boolean hasTitle;
