+++ WordPress/src/main/java/org/wordpress/android/ui/posts/PostsActivity.java
@@ -26,6 +26,7 @@ import org.wordpress.android.ui.RequestCodes;
 import org.wordpress.android.ui.posts.PostsListFragment.OnPostActionListener;
 import org.wordpress.android.ui.posts.PostsListFragment.OnPostSelectedListener;
 import org.wordpress.android.ui.posts.ViewPostFragment.OnDetailPostActionListener;
+import org.wordpress.android.util.AccountHelper;
 import org.wordpress.android.util.AlertUtils;
 import org.wordpress.android.util.AppLog;
 import org.wordpress.android.util.AppLog.T;
@@ -190,7 +191,7 @@ public class PostsActivity extends ActionBarActivity
 
         // posts can't be shown if there aren't any visible blogs, so redirect to the reader and
         // exit the post list in this situation
-        if (WordPress.isSignedIn()) {
+        if (AccountHelper.getDefaultAccount().isSignedIn()) {
             // TODO:
             /*if (showCorrectActivityForAccountIfRequired()) {
                 finish();
