+++ WordPress/src/main/java/org/wordpress/android/ui/prefs/SettingsFragment.java
@@ -39,6 +39,7 @@ import org.wordpress.android.ui.accounts.ManageBlogsActivity;
 import org.wordpress.android.ui.accounts.NewBlogActivity;
 import org.wordpress.android.ui.accounts.SignInActivity;
 import org.wordpress.android.ui.notifications.utils.NotificationsUtils;
+import org.wordpress.android.util.AccountHelper;
 import org.wordpress.android.util.ActivityUtils;
 import org.wordpress.android.util.AppLog;
 import org.wordpress.android.util.AppLog.T;
@@ -137,7 +138,7 @@ public class SettingsFragment extends PreferenceFragment {
 
     private void initNotifications() {
         // AuthenticatorRequest notification settings if needed
-        if (WordPress.hasDotComToken()) {
+        if (AccountHelper.getDefaultAccount().hasAccessToken()) {
             String settingsJson = mSettings.getString(NotificationsUtils.WPCOM_PUSH_DEVICE_NOTIFICATION_SETTINGS, null);
             if (settingsJson == null) {
                 com.wordpress.rest.RestRequest.Listener listener = new RestRequest.Listener() {
@@ -262,8 +263,8 @@ public class SettingsFragment extends PreferenceFragment {
     }
 
     private void addWpComSignIn(PreferenceCategory wpComCategory, int order) {
-        if (WordPress.hasDotComToken()) {
-            String username = mSettings.getString(WordPress.WPCOM_USERNAME_PREFERENCE, null);
+        if (AccountHelper.getDefaultAccount().hasAccessToken()) {
+            String username = AccountHelper.getDefaultAccount().getUserName();
             Preference usernamePref = new Preference(getActivity());
             usernamePref.setTitle(getString(R.string.username));
             usernamePref.setSummary(username);
