+++ WordPress/src/main/java/org/wordpress/android/ui/reader/views/ReaderWebView.java
@@ -2,9 +2,7 @@ package org.wordpress.android.ui.reader.views;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.content.SharedPreferences;
 import android.os.Build;
-import android.preference.PreferenceManager;
 import android.text.TextUtils;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
@@ -17,6 +15,7 @@ import android.webkit.WebView;
 import android.webkit.WebViewClient;
 
 import org.wordpress.android.WordPress;
+import org.wordpress.android.util.AccountHelper;
 import org.wordpress.android.util.AppLog;
 import org.wordpress.android.util.UrlUtils;
 import org.wordpress.android.util.WPRestClient;
@@ -62,7 +61,7 @@ public class ReaderWebView extends WebView {
 
     public ReaderWebView(Context context) {
         super(context);
-        init(context);
+        init();
     }
 
     @Override
@@ -77,19 +76,18 @@ public class ReaderWebView extends WebView {
 
     public ReaderWebView(Context context, AttributeSet attrs) {
         super(context, attrs);
-        init(context);
+        init();
     }
 
     public ReaderWebView(Context context, AttributeSet attrs, int defStyle) {
         super(context, attrs, defStyle);
-        init(context);
+        init();
     }
 
     @SuppressLint("NewApi")
-    private void init(Context context) {
+    private void init() {
         if (!isInEditMode()) {
-            SharedPreferences settings = PreferenceManager.getDefaultSharedPreferences(context);
-            mToken = settings.getString(WordPress.ACCESS_TOKEN_PREFERENCE, "");
+            mToken = AccountHelper.getDefaultAccount().getAccessToken();
 
             mReaderChromeClient = new ReaderWebChromeClient(this);
             this.setWebChromeClient(mReaderChromeClient);
