+++ WordPress/src/main/java/org/wordpress/android/ui/stats/StatsWPLinkMovementMethod.java
@@ -1,15 +1,13 @@
 package org.wordpress.android.ui.stats;
 
-import android.content.SharedPreferences;
-import android.preference.PreferenceManager;
 import android.text.Layout;
 import android.text.Spannable;
 import android.text.style.URLSpan;
 import android.view.MotionEvent;
 import android.widget.TextView;
 
-import org.wordpress.android.WordPress;
 import org.wordpress.android.ui.WPWebViewActivity;
+import org.wordpress.android.util.AccountHelper;
 import org.wordpress.android.util.AppLog;
 import org.wordpress.android.util.UrlUtils;
 import org.wordpress.android.util.WPLinkMovementMethod;
@@ -61,8 +59,7 @@ public class StatsWPLinkMovementMethod extends WPLinkMovementMethod {
                     }
                     AppLog.d(AppLog.T.UTILS, "Opening the Authenticated in-app browser : " + url);
                     // Let's try the global wpcom credentials
-                    SharedPreferences settings = PreferenceManager.getDefaultSharedPreferences(widget.getContext());
-                    String statsAuthenticatedUser = settings.getString(WordPress.WPCOM_USERNAME_PREFERENCE, null);
+                    String statsAuthenticatedUser = AccountHelper.getDefaultAccount().getUserName();
                     if (org.apache.commons.lang.StringUtils.isEmpty(statsAuthenticatedUser)) {
                         // Still empty. Do not eat the event, but let's open the default Web Browser.
                         return super.onTouchEvent(widget, buffer, event);
