+++ WordPress/src/main/java/org/wordpress/android/ui/themes/ThemePreviewFragment.java
@@ -17,6 +17,7 @@ import org.wordpress.android.R;
 import org.wordpress.android.WordPress;
 import org.wordpress.android.models.Blog;
 import org.wordpress.android.ui.WPWebViewActivity;
+import org.wordpress.android.util.AccountHelper;
 import org.wordpress.android.util.helpers.WPWebChromeClient;
 import org.wordpress.android.util.WPWebViewClient;
 
@@ -44,7 +45,6 @@ public class ThemePreviewFragment extends Fragment {
         public void onPause(Fragment fragment);
         public void onActivateThemeClicked(String themeId, Fragment fragment);
     }
-
 
     public static ThemePreviewFragment newInstance(String themeId, String previewURL) {
         ThemePreviewFragment fragment = new ThemePreviewFragment();
@@ -149,7 +149,7 @@ public class ThemePreviewFragment extends Fragment {
     public void onActivityCreated(Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
 
-        mWebView.setWebViewClient(new WPWebViewClient(getActivity(), mBlog));
+        mWebView.setWebViewClient(new WPWebViewClient(mBlog));
     }
 
     private void refreshViews() {
@@ -169,7 +169,7 @@ public class ThemePreviewFragment extends Fragment {
 
         String authenticationUrl = WordPress.getLoginUrl(mBlog);
         String postData = WPWebViewActivity.getAuthenticationPostData(authenticationUrl, url,
-                mBlog.getUsername(), mBlog.getPassword(), WordPress.getDotComToken()
+                mBlog.getUsername(), mBlog.getPassword(), AccountHelper.getDefaultAccount().getAccessToken()
         );
 
         mWebView.postUrl(authenticationUrl, postData.getBytes());
