--- WordPress/src/main/java/org/wordpress/android/util/AccountActions.java
@@ -1,38 +1,0 @@
-package org.wordpress.android.util;
-
-import com.android.volley.VolleyError;
-import com.wordpress.rest.RestRequest;
-
-import org.json.JSONObject;
-import org.wordpress.android.WordPress;
-import org.wordpress.android.datasets.AccountTable;
-import org.wordpress.android.models.Account;
-import org.wordpress.android.util.AppLog.T;
-
-public class AccountActions {
-    public static void fecthAccount(final Account account) {
-        com.wordpress.rest.RestRequest.Listener listener = new RestRequest.Listener() {
-            @Override
-            public void onResponse(JSONObject jsonObject) {
-                if (jsonObject != null) {
-                    account.updateFromRestResponse(jsonObject);
-                    AccountTable.save(account);
-                }
-            }
-        };
-
-        RestRequest.ErrorListener errorListener = new RestRequest.ErrorListener() {
-            @Override
-            public void onErrorResponse(VolleyError volleyError) {
-                AppLog.e(T.API, volleyError);
-            }
-        };
-
-        WordPress.getRestClientUtilsV1_1().get("me", listener, errorListener);
-    }
-
-    public static void signout(Account account) {
-        account.init();
-        AccountTable.save(account);
-    }
-}
