+++ WordPress/src/main/java/org/xmlrpc/android/XMLRPCClient.java
@@ -20,6 +20,7 @@ import org.apache.http.params.HttpParams;
 import org.apache.http.params.HttpProtocolParams;
 import org.apache.http.util.EntityUtils;
 import org.wordpress.android.WordPress;
+import org.wordpress.android.util.AccountHelper;
 import org.wordpress.android.util.AppLog;
 import org.wordpress.android.util.AppLog.T;
 import org.wordpress.android.util.CoreEvents;
@@ -623,7 +624,7 @@ public class XMLRPCClient implements XMLRPCClientInterface {
         if (ctx == null) return;
 
         if (isDotComXMLRPCEndpoint(mPostMethod.getURI())) {
-            String token = WordPress.getDotComToken();
+            String token = AccountHelper.getDefaultAccount().getAccessToken();
             if (!TextUtils.isEmpty(token)) {
                 setAuthorizationHeader(token);
             }
